wave fixes

+++ convolution/Assets/Scripts/waves/Boxcar.cs
@@ -4,7 +4,7 @@ using System;
 
 public class Boxcar : AbstractWave {
 
-    public override float get(int x) {
+    public override float get(float x) {
         return (x >= 0 && x <= 1) ? 1.0f : 0.0f;
     }
 }
