+++ convolution/Assets/Scripts/waves/Sine.cs
@@ -4,7 +4,7 @@ using System;
 
 public class Sine : AbstractWave {
 
-    public override float get(int i) {
+    public override float get(float i) {
         return Mathf.Sin(i * (_isReversed ? -1 : 1) * _freq) * _amp;
     }
 }
